fix(backend): resolve startup blockers and lazy-load bugs
- Fix Windows mvnw.cmd trailing backslash corrupting Maven args
- Map Docker postgres to port 5433 to avoid conflict with local PG18
- Add ?TimeZone=UTC&stringtype=unspecified to JDBC URL: fixes Asia/Calcutta
  rejection by PG16 and implicit cast of trade_direction enum strings
- Change risk_tolerance entity field to short (matches SMALLINT in schema)
- Add @Transactional(readOnly=true) to ActivityReportService and
  TradeService#getTradeHistory to prevent LazyInitializationException
  when traversing ClientTrade->ClientOrder->Instrument lazy associations
- Gitignore .env and application-local.properties from backend

+++ backend-trade-nexus/src/main/java/com/marshals/model/ClientPreferences.java
@@ -26,7 +26,7 @@ public class ClientPreferences {
     private String percentageOfSpend;
 
     @Column(name = "risk_tolerance", nullable = false)
-    private int riskTolerance;
+    private short riskTolerance;
 
     @Column(name = "accept_advisor", nullable = false)
     private boolean acceptAdvisor;
@@ -48,8 +48,8 @@ public class ClientPreferences {
     public String getPercentageOfSpend() { return percentageOfSpend; }
     public void setPercentageOfSpend(String percentageOfSpend) { this.percentageOfSpend = percentageOfSpend; }
 
-    public int getRiskTolerance() { return riskTolerance; }
-    public void setRiskTolerance(int riskTolerance) { this.riskTolerance = riskTolerance; }
+    public short getRiskTolerance() { return riskTolerance; }
+    public void setRiskTolerance(short riskTolerance) { this.riskTolerance = riskTolerance; }
 
     public boolean isAcceptAdvisor() { return acceptAdvisor; }
     public void setAcceptAdvisor(boolean acceptAdvisor) { this.acceptAdvisor = acceptAdvisor; }
